tests/requirements.txt
jinja2>=3.0
pytest>=7.0
pyyaml>=6.0
